pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "napsack"
version = "0.1.0"
readme = "README.md"
description = "NAPsack records and aggregates your computer use — screenshots plus input events (click, keypress, scroll, cursor move). It groups activity into event bursts and uses a VLM pipeline to generate human-readable captions describing what happened."
requires-python = ">=3.11,<=3.13"
dependencies = [
    "mss (>=10.0.0,<11.0.0)",
    "numpy==2.2",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "ruptures (>=1.1.9,<2.0.0)",
    "ipdb>=0.13.13,<0.14.0",
    "plotly (>=6.2.0,<7.0.0)",
    "nbformat (>=5.10.4,<6.0.0)",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.5",
    "datasets>=3.0.0,<4.0.0",
    "pillow>=11.3.0",
    "imageio>=2.37.0",
    "screeninfo>=0.8.1",
    "pynput>=1.8.1",
    "scikit-image>=0.25.2",
    "google-genai>=1.45.0",
    "openai>=2.8.1",
    "google-cloud-storage>=3.6.0",
    "google-cloud-bigquery>=3.38.0",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
label = ["prompts/*.txt"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["label*", "record*"]
